+++ hypha/apply/funds/views_beta.py
@@ -8,14 +8,11 @@ from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render
 from django.utils.translation import gettext as _
 from django_tables2 import RequestConfig
-from wagtail.models import Page
 
 from hypha.apply.funds.workflow import PHASES
-from hypha.apply.users.groups import STAFF_GROUP_NAME
 
 from .models import (
     ApplicationSubmission,
-    Round,
 )
 from .permissions import is_user_has_access_to_view_archived_submissions
 from .tables import (
@@ -161,98 +158,3 @@ def submission_dashboard(request: HttpRequest, template_name='submissions/all.ht
 
 # submission_all_beta = SubmissionAdminListView.as_view()
 submission_all_beta = submission_dashboard
-
-
-def sub_menu_funds(request):
-    selected_funds = request.GET.getlist("fund")
-
-    # Funds Filter Options
-    funds = [{
-        'id': f.id,
-        'selected': str(f.id) in selected_funds,
-        'title': f.title
-    } for f in Page.objects.filter(
-        applicationsubmission__isnull=False
-    ).order_by("title").distinct()]
-
-    ctx = {
-        'funds': funds,
-        'selected_funds': selected_funds,
-    }
-
-    return render(request, "submissions/submenu/funds.html", ctx)
-
-
-def sub_menu_leads(request):
-    selected_leads = request.GET.getlist("lead")
-
-    leads = [{
-        'id': item.id,
-        'selected': str(item.id) in selected_leads,
-        'title': str(item),
-        'slack': item.slack,
-    } for item in User.objects.filter(
-        submission_lead__isnull=False
-    ).order_by().distinct()]
-
-    ctx = {
-        'leads': leads,
-        'selected_leads': selected_leads,
-    }
-
-    return render(request, "submissions/submenu/leads.html", ctx)
-
-
-def sub_menu_rounds(request):
-    selected_rounds = request.GET.getlist("round")
-    selected_fund = request.GET.get("fund")
-    qs = Round.objects.all()
-
-    if selected_fund:
-        fund = Page.objects.get(id=selected_fund)
-        qs = Round.objects.child_of(fund)
-
-    open_rounds = [{
-        'id': item.id,
-        'selected': str(item.id) in selected_rounds,
-        'title': str(item)
-    } for item in qs.open().order_by('-end_date').distinct()]
-
-    closed_rounds = [{
-        'id': item.id,
-        'selected': str(item.id) in selected_rounds,
-        'title': str(item)
-    } for item in qs.closed().filter(
-        submissions__isnull=False
-    ).order_by('-end_date').distinct()]
-
-    ctx = {
-        'open_rounds': open_rounds,
-        'closed_rounds': closed_rounds,
-        'selected_rounds': selected_rounds,
-    }
-
-    return render(request, "submissions/submenu/rounds.html", ctx)
-
-
-def sub_menu_reviewers(request):
-    selected_reviewers = request.GET.getlist("reviewers")
-
-    reviewers = [{
-        'id': item.id,
-        'selected': str(item.id) in selected_reviewers,
-        'title': str(item),
-        'slack': item.slack,
-    } for item in User.objects.filter(
-        models.Q(submissions_reviewer__isnull=False) | models.Q(groups__name=STAFF_GROUP_NAME) | models.Q(is_superuser=True)
-    ).order_by().distinct()]
-
-
-    reviewers = sorted(reviewers, key=lambda t: t['selected'], reverse=True)
-
-    ctx = {
-        'reviewers': reviewers,
-        'selected_reviewers': selected_reviewers,
-    }
-
-    return render(request, "submissions/submenu/reviewers.html", ctx)
