+++ hypha/apply/users/templates/two_factor/core/backup_tokens.html
@@ -53,8 +53,6 @@
 
 {% block extra_js %}
 <script src="{% static 'js/clipboard.min.js' %}"></script>
-<script src="{% static 'js/tooltips/popper.min.js' %}"></script>
-<script src="{% static 'js/tooltips/tippy.min.js' %}"></script>
 <script>
 
     {# Select the codes when clicked #}
